pyproject.toml
[project]
name = "video-subtitle-extractor"
version = "0.1.0"
description = "Extract audio from video and generate subtitles using Whisper"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "openai-whisper>=20231117",
    "moviepy>=1.0.3",
    "pydub>=0.25.1",
    "ffmpeg-python>=0.2.0",
    "tqdm>=4.66.0",
    "srt>=3.5.3",
    "pytube>=15.0.0",
    "google-api-python-client>=2.100.0",
    "opencv-python>=4.8.0",
    "easyocr>=1.7.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
]

[tool.uv]
dev-dependencies = [
    "ipython>=8.0.0",
]
